+++ Src/DiscInfo.cpp
@@ -1,6 +1,6 @@
 /****************************************************************
 BeebEm - BBC Micro and Master 128 Emulator
-Copyright (C) 2023 Chris Needham
+Copyright (C) 2009  Mike Wyatt
 
 This program is free software; you can redistribute it and/or
 modify it under the terms of the GNU General Public License
@@ -12,22 +12,32 @@ but WITHOUT ANY WARRANTY; without even the implied warranty of
 MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 GNU General Public License for more details.
 
-You should have received a copy of the GNU General Public
-License along with this program; if not, write to the Free
+You should have received a copy of the GNU General Public 
+License along with this program; if not, write to the Free 
 Software Foundation, Inc., 51 Franklin Street, Fifth Floor,
 Boston, MA  02110-1301, USA.
 ****************************************************************/
 
-#ifndef FILE_UTILS_HEADER
-#define FILE_UTILS_HEADER
+#include <stdio.h>
 
-#include <string>
+#include "DiscInfo.h"
+#include "DiscEdit.h"
+#include "FileUtils.h"
 
-bool FileExists(const char* PathName);
-bool FolderExists(const char* PathName);
+DiscInfoType DiscInfo[2];
 
-std::string AppendPath(const std::string& BasePath, const std::string& Path);
+bool IsDoubleSidedSSD(const char *FileName, FILE *pFile)
+{
+	if (!HasFileExt(FileName, ".ssd"))
+	{
+		return false;
+	}
 
-bool HasFileExt(const char* FileName, const char* Ext);
+	fseek(pFile, 0, SEEK_END);
 
-#endif
+	long Size = ftell(pFile);
+
+	fseek(pFile, 0, SEEK_SET);
+
+	return Size > 80 * DFS_SECTORS_PER_TRACK * DFS_SECTOR_SIZE;
+}